Bug 1341102: Fix bustage on a CLOSED TREE; r=bustage
MozReview-Commit-ID: Ho2PNnBhR3M

+++ gfx/src/nsCoord.h
@@ -132,8 +132,8 @@ inline nscoord _nscoordSaturatingMultiply(nscoord aCoord, float aScale,
 #else
   float product = aCoord * aScale;
   if (requireNotNegative ? aCoord > 0 : (aCoord > 0) == (aScale > 0))
-    return NSToCoordRoundWithClamp(std::min<float>(nscoord_MAX, product));
-  return NSToCoordRoundWithClamp(std::max<float>(nscoord_MIN, product));
+    return NSToCoordRoundWithClamp(std::min<float>((float)nscoord_MAX, product));
+  return NSToCoordRoundWithClamp(std::max<float>((float)nscoord_MIN, product));
 #endif
 }
